firmware_loader: Add /lib/firmware/vendor paths for Apple ARM platforms
On Apple ARM machines (branded Apple Silicon Macs), there is
non-redistributable firmware provided in macOS that is required
for Linux to leverage the hardware to run properly.

The firmware covers Wi-Fi+BT support, USB, hardware ISP, the touchbar
touchscreen, and the trackpad.

In order to ensure this firmware can be installed without messing
with firmware provided by linux-firmware and other packages that
put things in the tree, the empty "vendor" subdirectory was created
to serve this purpose. This directory is used by the Asahi Linux
installation process to transfer firmware from macOS to the newly
installed Linux system.

This is based on the patch originally written by Hector Martin
for Asahi Linux.

Link: https://asahilinux.org/docs/platform/introduction/

Signed-off-by: Neal Gompa <neal@gompa.dev>